requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==12.0
redis[hiredis]==5.0.0
langchain==0.3.0
langchain-openai==0.2.0
openai==1.45.0
httpx==0.27.0
asyncpg==0.29.0
celery[redis]==5.4.0
twilio==9.0.0
python-dotenv==1.0.1
pydantic==2.8.0
